% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py.assign.R
\name{py.assign}
\alias{py.assign}
\title{Assign and get variables in Python from R}
\usage{
py.assign(var.name, value,
  json.opt.args = getOption("SnakeCharmR.json.opt.args", list()))
}
\arguments{
\item{var.name}{a character string containing a valid python variable name}

\item{value}{an R object whose equivalent wants to be assigned to the
variable in python}

\item{json.opt.args}{explicit arguments to pass for JSON transformation}

\item{...}{other arguments passed to the internal \code{toJSON} function
call}
}
\value{
Function \code{py.get} returns a R version of the Python
variable \code{py.var}.
}
\description{
Functions that assign and get Python variables from R.
}
\details{
These functions can assign values to variables in Python as well as get
their values back to R.  Objects are serialized as json strings while being
transferred between R and Python.
}
\examples{
a <- 1:4
py.assign( "a", a )
py.exec( "b = len( a )" )
py.get( "b" )

py.exec( "import math" )
py.get( "math.pi" )
}
\references{
\url{http://code.google.com/p/simplejson}
}
\keyword{manip}

