% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py.load.R
\name{py.load}
\alias{py.load}
\title{Executes Python code.}
\usage{
py.load(file, get.exception = TRUE)
}
\arguments{
\item{file}{a file containing python code to be executed}

\item{get.exception}{logical value indicating whether to check or not for
exceptions in Python}
}
\value{
None.  If the code produces some output, it is up to the caller to
go and fetch if from Python using function \code{\link{py.get}}.
}
\description{
This function runs Python code contained in a file.  Typically, this file
would contain functions to be called via \code{\link{py.call}} or other
functions in this package.
}
\details{
The \code{get.exception} option allows the user to disregard Python
exceptions in cases where safe calls to avoid the overhead of checking for
them.
}
\examples{
a <- 1:4
b <- 5:8

# this file contains the definition of function concat
py.load( system.file( "concat.py", package = "rPython" ) )
py.call( "concat", a, b)
}
\keyword{manip}

