% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnakeCharmR.r
\docType{package}
\name{SnakeCharmR}
\alias{SnakeCharmR}
\alias{SnakeCharmR-package}
\title{R and Python Integration}
\description{
A fork from rPython which uses jsonlite and has several fixes and improvements
}
\author{
Alexandre Sieira
}
