% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py.rm.R
\name{py.rm}
\alias{py.rm}
\title{Remove a python variable}
\usage{
py.rm(var.name)
}
\arguments{
\item{var.name}{a character string containing a valid python variable name}
}
\value{
invisible result of python `del` operation.
}
\description{
Remove a python variable
}
\keyword{manip}

